feat(label): advanced configs only, and & or
