+++ packages/release-mastermind/.github/allconfigs.json
@@ -2038,14 +2038,18 @@
               "sections": [
                 {
                   "title": "Features Effectuated",
-                  "PRlabels": ["Type - Feature"],
+                  "PRlabels": [
+                    "Type - Feature"
+                  ],
                   "issueLabels": [],
                   "includeCommitter": true,
                   "linkPR": true
                 },
                 {
                   "title": "Bugs Squashed",
-                  "PRlabels": ["Type - Fix"],
+                  "PRlabels": [
+                    "Type - Fix"
+                  ],
                   "issueLabels": [],
                   "includeCommitter": true,
                   "linkPR": true
@@ -2078,7 +2082,9 @@
                 {
                   "title": "Confession Time",
                   "PRlabels": [],
-                  "issueLabels": ["Bug - Confirmed"],
+                  "issueLabels": [
+                    "Bug - Confirmed"
+                  ],
                   "includeCommitter": true,
                   "linkPR": true
                 }
@@ -2107,7 +2113,9 @@
             "conventions": [
               {
                 "requires": 1,
-                "conditions": [{}]
+                "conditions": [
+                  {}
+                ]
               }
             ]
           },
@@ -2463,7 +2471,10 @@
         "project": {
           "ref": "develop",
           "enforceConventions": {
-            "onColumn": ["Accepted", "Reviewing"],
+            "onColumn": [
+              "Accepted",
+              "Reviewing"
+            ],
             "conventions": [
               {
                 "requires": 1,
@@ -2516,7 +2527,7 @@
                   "type": "onColumn",
                   "project": "Developer Operations",
                   "column": "In Progress"
-                },
+                }, 
                 {
                   "type": "onColumn",
                   "project": "Developer Operations",
