+++ packages/release-mastermind/src/conditions/util/index.ts
@@ -1,9 +1,11 @@
+import and, { ConditionAnd } from './and'
 import creatorMatches, { ConditionCreatorMatches } from './creatorMatches'
 import descriptionMatches, {
   ConditionDescriptionMatches
 } from './descriptionMatches'
 import hasLabel, { ConditionHasLabel } from './hasLabel'
 import isOpen, { ConditionIsOpen } from './isOpen'
+import or, { ConditionOr } from './or'
 import titleMatches, { ConditionTitleMatches } from './titleMatches'
 
 export type Condition =
@@ -12,11 +14,15 @@ export type Condition =
   | ConditionIsOpen
   | ConditionTitleMatches
   | ConditionHasLabel
+  | ConditionOr
+  | ConditionAnd
 
 export const handlers = [
   creatorMatches,
   descriptionMatches,
   isOpen,
   hasLabel,
-  titleMatches
+  titleMatches,
+  and,
+  or
 ]
