+++ packages/release-mastermind/src/releaseMastermind.ts
@@ -7,7 +7,7 @@ import { log } from '.'
 import { CurContext } from './conditions'
 import { contextHandler } from './contextHandler'
 import { Issues, Project, PullRequests } from './contexts'
-import { Config, Label, Options, Runners } from './types'
+import { Config, Label, Options, Runners } from '../types'
 import { utils } from './utils'
 
 let local: any
